corrected type received by ft_printf_xX

+++ ft_printf_pdix.c
@@ -6,7 +6,7 @@
 /*   By: sde-alva <sde-alva@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/10 18:12:09 by sde-alva          #+#    #+#             */
-/*   Updated: 2021/08/18 12:06:53 by sde-alva         ###   ########.fr       */
+/*   Updated: 2021/08/18 14:21:23 by sde-alva         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ int	ft_printf_u(unsigned int un, t_printf_flags *flags)
 	return (n_printed);
 }
 
-int	ft_printf_xX(unsigned long n, t_printf_flags *flags, char upper)
+int	ft_printf_xX(unsigned long long n, t_printf_flags *flags, char upper)
 {
 	int		n_printed;
 	char	*x_num;
